+++ src/main/java/com/example/outsourcing/user/service/UserService.java
@@ -1,7 +1,5 @@
 package com.example.outsourcing.user.service;
 
-import com.example.outsourcing.address.dto.UpdateUserAddressRequestDTO;
-import com.example.outsourcing.address.entity.Address;
 import com.example.outsourcing.common.config.PasswordEncoder;
 import com.example.outsourcing.image.service.ImageService;
 import com.example.outsourcing.user.dto.PwdUpdateRequestDTO;
@@ -12,8 +10,6 @@ import com.example.outsourcing.user.dto.UserUpdateRequestDTO;
 import com.example.outsourcing.user.entity.User;
 import com.example.outsourcing.user.repository.UserRepository;
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -124,142 +120,5 @@ public class UserService {
     } else {
       return null;
     }
-  }
-
-  // 유저 주소 등록
-  @Transactional
-  public void createAddress(Long userId, UpdateUserAddressRequestDTO requestDTO) {
-    User user = userRepository.findById(userId).orElseThrow(
-        () -> new RuntimeException("유저를 찾을 수 없습니다."));
-
-    // 최대 주소(3개) 인지 검사
-    if (user.hasMaxAddresses()) {
-      throw new RuntimeException("주소는 최대 3개 등록 가능합니다.");
-    }
-
-    Address address = new Address(requestDTO, user);
-
-    // 주소를 가지지 않거나, 새로 추가하면서 디폴트 설정을 해주고 싶다면
-    if (user.getAddresses().isEmpty() || requestDTO.isDefault()) {
-      address.setDefault(true);
-      user.getAddresses().add(address);
-      return;
-    }
-
-    // 기존 주소가 있고, 디폴트 주소를 설정하고 싶다면
-    if (requestDTO.isDefault()) {
-      for (Address nowAddress : user.getAddresses()) {
-        // 디폴트 주소 획득
-        if (nowAddress.isDefault()) {
-          nowAddress.setDefault(false);
-        }
-      }
-    }
-
-    user.getAddresses().add(address);
-  }
-
-  // 유저 주소 업데이트
-  @Transactional
-  public void updateAddress(Long userId, Long addressId,
-      UpdateUserAddressRequestDTO requestDTO) {
-    User user = userRepository.findById(userId).orElseThrow(
-        () -> new RuntimeException("유저를 찾을 수 없습니다."));
-
-    Address defaultAddress = null;
-    Address targetAddress = null;
-
-    for (Address address : user.getAddresses()) {
-      // 디폴트 주소 획득
-      if (defaultAddress == null && address.isDefault()) {
-        defaultAddress = address;
-      }
-      // 디폴트 설정 대상 주소
-      if (targetAddress == null && Objects.equals(address.getId(), addressId)) {
-        targetAddress = address;
-      }
-    }
-
-    // 디폴트 설정 대상 주소 조회 성공이라면
-    if (targetAddress != null) {
-      targetAddress.setAddress(requestDTO.getAddress());
-
-      // 디폴트 주소 설정 한다면
-      if (requestDTO.isDefault()) {
-        targetAddress.setDefault(true);
-
-        // 디폴트 주소가 있다면
-        if (defaultAddress != null) {
-          defaultAddress.setDefault(false);
-        }
-      }
-      return;
-    }
-
-    throw new RuntimeException("사용자의 주소가 아닙니다.");
-  }
-
-  // 유저 주소 디폴트 설정
-  @Transactional
-  public void setDefaultAddress(Long userId, Long addressId) {
-    User user = userRepository.findById(userId).orElseThrow(
-        () -> new RuntimeException("유저를 찾을 수 없습니다."));
-
-    Address defaultAddress = null;
-    Address targetAddress = null;
-
-    for (Address address : user.getAddresses()) {
-      // 디폴트 주소 획득
-      if (defaultAddress == null && address.isDefault()) {
-        defaultAddress = address;
-      }
-      // 디폴트 설정 대상 주소
-      if (targetAddress == null && Objects.equals(address.getId(), addressId)) {
-        targetAddress = address;
-      }
-    }
-
-    // 디폴트 설정 대상 주소 조회 성공이라면
-    if (targetAddress != null) {
-      targetAddress.setDefault(true);
-      // 디폴트 주소가 있다면
-      if (defaultAddress != null) {
-        defaultAddress.setDefault(false);
-      }
-      return;
-    }
-
-    throw new RuntimeException("사용자의 주소가 아닙니다.");
-  }
-
-  @Transactional
-  public void deleteAddress(Long userId, Long addressId) {
-    User user = userRepository.findById(userId).orElseThrow(
-        () -> new RuntimeException("유저를 찾을 수 없습니다."));
-    List<Address> addresses = user.getAddresses();
-    Address defaultAddress = null;
-    Address targetAddress = null;
-
-    for (Address address : user.getAddresses()) {
-      // 디폴트 주소 획득
-      if (defaultAddress == null && address.isDefault()) {
-        defaultAddress = address;
-      }
-      // 삭제 대상 주소
-      if (targetAddress == null && Objects.equals(address.getId(), addressId)) {
-        targetAddress = address;
-      }
-      // 디폴트, 삭제 주소 모두 획득
-      if (targetAddress != null && defaultAddress != null) {
-        if (Objects.equals(defaultAddress.getId(), targetAddress.getId())) {
-          addresses.remove(address);
-          addresses.get(addresses.size() - 1).setDefault(true);
-          return;
-        }
-        addresses.remove(address);
-        return;
-      }
-    }
-    throw new RuntimeException("사용자의 주소가 아닙니다.");
   }
 }
